Include Manifest.txt and lib/kryptcore.jar in the repo.
This is needed for CI for the moment because krypt requires krypt-core-jruby as a gem, and the 'gem build' command will actually assume both files to be already present. Therefore we need to commit them to the repo, too, because they won't be built on the fly when required by krypt.